+++ components/map/MapAutocomplete.tsx
@@ -16,9 +16,10 @@ import FlexRow from '@/components/shared/FlexRow'
 import Icon from '@/components/shared/Icon'
 import Typography from '@/components/shared/Typography'
 import { useLocale, useTranslation } from '@/hooks/useTranslation'
+import { GeocodingFeature } from '@/modules/arcgis/types'
 import { useLocation } from '@/modules/map/hooks/useLocation'
 import { useMapAutocompleteGetOptions } from '@/modules/map/hooks/useMapAutocompleteGetOptions'
-import { GeocodingFeature, isGeocodingFeature, UdrZoneFeature } from '@/modules/map/types'
+import { isGeocodingFeature, UdrZoneFeature } from '@/modules/map/types'
 import { findShapesInRadius } from '@/modules/map/utils/findShapesInRadius'
 import { forwardGeocode } from '@/modules/map/utils/forwardGeocode'
 import { translateMapObject } from '@/modules/map/utils/translateMapObject'
